moved cmdb graph to connectors
